+++ opentelemetry-sdk/tests/metrics/integration_test/test_explicit_bucket_histogram_aggregation.py
@@ -18,6 +18,7 @@ from unittest import TestCase
 
 from pytest import mark
 
+from opentelemetry.metrics import NoOpMeterProvider
 from opentelemetry.sdk.metrics import Histogram, MeterProvider
 from opentelemetry.sdk.metrics.export import (
     AggregationTemporality,
@@ -46,6 +47,9 @@ class TestExplicitBucketHistogramAggregation(TestCase):
         )
 
         provider = MeterProvider(metric_readers=[reader])
+        # Disable SDK metrics
+        # pylint: disable=protected-access
+        reader._set_meter_provider(NoOpMeterProvider())
         meter = provider.get_meter("name", "version")
 
         histogram = meter.create_histogram("histogram")
@@ -159,7 +163,7 @@ class TestExplicitBucketHistogramAggregation(TestCase):
         provider.shutdown()
 
     @mark.skipif(
-        system() != "Linux",
+        system() == "Windows",
         reason=(
             "Tests fail because Windows time_ns resolution is too low so "
             "two different time measurements may end up having the exact same"
@@ -177,6 +181,9 @@ class TestExplicitBucketHistogramAggregation(TestCase):
         )
 
         provider = MeterProvider(metric_readers=[reader])
+        # Disable SDK metrics
+        # pylint: disable=protected-access
+        reader._set_meter_provider(NoOpMeterProvider())
         meter = provider.get_meter("name", "version")
 
         histogram = meter.create_histogram("histogram")