+++ opentelemetry-sdk/tests/metrics/integration_test/test_sum_aggregation.py
@@ -21,7 +21,7 @@ from unittest import TestCase
 from pytest import mark
 
 from opentelemetry.context import Context
-from opentelemetry.metrics import Observation
+from opentelemetry.metrics import NoOpMeterProvider, Observation
 from opentelemetry.sdk.metrics import Counter, MeterProvider, ObservableCounter
 from opentelemetry.sdk.metrics._internal.exemplar import AlwaysOnExemplarFilter
 from opentelemetry.sdk.metrics.export import (
@@ -33,7 +33,7 @@ from opentelemetry.sdk.metrics.view import SumAggregation
 
 class TestSumAggregation(TestCase):
     @mark.skipif(
-        system() != "Linux",
+        system() == "Windows",
         reason=(
             "Tests fail because Windows time_ns resolution is too low so "
             "two different time measurements may end up having the exact same"
@@ -68,6 +68,9 @@ class TestSumAggregation(TestCase):
         )
 
         provider = MeterProvider(metric_readers=[reader])
+        # Disable SDK metrics
+        # pylint: disable=protected-access
+        reader._set_meter_provider(NoOpMeterProvider())
         meter = provider.get_meter("name", "version")
 
         meter.create_observable_counter(
@@ -156,7 +159,7 @@ class TestSumAggregation(TestCase):
             self.assertIsNone(metrics_data)
 
     @mark.skipif(
-        system() != "Linux",
+        system() == "Windows",
         reason=(
             "Tests fail because Windows time_ns resolution is too low so "
             "two different time measurements may end up having the exact same"
@@ -191,6 +194,9 @@ class TestSumAggregation(TestCase):
         )
 
         provider = MeterProvider(metric_readers=[reader])
+        # Disable SDK metrics
+        # pylint: disable=protected-access
+        reader._set_meter_provider(NoOpMeterProvider())
         meter = provider.get_meter("name", "version")
 
         meter.create_observable_counter(
@@ -251,7 +257,7 @@ class TestSumAggregation(TestCase):
             self.assertIsNone(metrics_data)
 
     @mark.skipif(
-        system() != "Linux",
+        system() == "Windows",
         reason=(
             "Tests fail because Windows time_ns resolution is too low so "
             "two different time measurements may end up having the exact same"
@@ -267,6 +273,9 @@ class TestSumAggregation(TestCase):
         )
 
         provider = MeterProvider(metric_readers=[reader])
+        # Disable SDK metrics
+        # pylint: disable=protected-access
+        reader._set_meter_provider(NoOpMeterProvider())
         meter = provider.get_meter("name", "version")
 
         counter = meter.create_counter("counter")
@@ -378,7 +387,7 @@ class TestSumAggregation(TestCase):
         provider.shutdown()
 
     @mark.skipif(
-        system() != "Linux",
+        system() == "Windows",
         reason=(
             "Tests fail because Windows time_ns resolution is too low so "
             "two different time measurements may end up having the exact same"
@@ -394,6 +403,9 @@ class TestSumAggregation(TestCase):
         )
 
         provider = MeterProvider(metric_readers=[reader])
+        # Disable SDK metrics
+        # pylint: disable=protected-access
+        reader._set_meter_provider(NoOpMeterProvider())
         meter = provider.get_meter("name", "version")
 
         counter = meter.create_counter("counter")